Fix code review issues: OBO auth propagation, dead code, model placement
- Add run_in_context wrapper to /benchmark/compare endpoint to propagate OBO ContextVar into thread pool
- Remove unreachable `return prompt` in result_comparator that referenced undefined variable
- Move CompareResultsRequest from router to models.py per project convention
- Move llm_utils imports to module level in result_comparator
- Remove dead selectAllByPriority/deselectAllByPriority exports from useAnalysis
